pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["namJeongwan <njw1107@naver.com>"]
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
requests = "^2.28.2"
pillow = "^9.5.0"
tox = "^4.4.12"
flask = "^2.2.3"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

